Сделал разгарничение удаления пользователей
Именно:
+ Админ может удалять только пользователей, младше его по рангу, как и регистрировать только научных руководителей и заведующего кафедрой
+ Рут может все, что и админ, а также регистрировать/удалять других админов

+++ src/main/java/g_server/g_server/application/controller/users/crud/UsersController.java
@@ -2,11 +2,16 @@ package g_server.g_server.application.controller.users.crud;
 
 import g_server.g_server.application.entity.users.Users;
 import g_server.g_server.application.repository.users.UsersRepository;
+import g_server.g_server.application.service.documents.DocumentUploadService;
 import g_server.g_server.application.service.users.UsersService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Optional;
+
+import static org.springframework.util.StringUtils.hasText;
 
 // Данный контроллер предназначен для администрирования пользователей админом
 // Смена ролей пользователей даже здесь не допускается
@@ -14,11 +19,16 @@ import java.util.Optional;
 // В дальнейшем надо реализовать возможность корректироки данных студентов и научных руководителей здесь
 @RestController
 public class UsersController {
+    public static final String AUTHORIZATION = "Authorization";
+
     @Autowired
     private UsersService usersService;
 
     @Autowired
     private UsersRepository usersRepository;
+
+    @Autowired
+    private DocumentUploadService documentUploadService;
 
     @GetMapping("/admin/users/all")
     public List<Users> findAll() {
@@ -31,11 +41,18 @@ public class UsersController {
     }
 
     @DeleteMapping("/admin/users/delete/{id}")
-    public String delete(@PathVariable int id) {
-        // Доделать чтобы админ не мог удалять себя и других админов
-        // Например ввести надуровень root-admin
-        usersRepository.deleteById(id);
-        return "Пользователь удален успешно";
+    public String delete(
+            @PathVariable int id,
+            HttpServletRequest httpServletRequest) {
+        String token = getTokenFromRequest(httpServletRequest);
+        Integer adminId = documentUploadService.getCreatorId(token);
+        if (usersService.checkUsersRoles(adminId, id)) {
+            usersRepository.deleteById(id);
+            return "Пользователь удален успешно";
+        }
+        else {
+            return "Недостаточно прав для удаления данного пользователя";
+        }
     }
 
     @PutMapping("/admin/users/change_email/")
@@ -51,5 +68,13 @@ public class UsersController {
         else {
             return "Такого пользователя не существует";
         }
+    }
+
+    private String getTokenFromRequest(HttpServletRequest request) {
+        String bearer = request.getHeader(AUTHORIZATION);
+        if (hasText(bearer) && bearer.startsWith("Bearer ")) {
+            return bearer.substring(7);
+        }
+        return null;
     }
 }
